requirements.txt
django==1.11.17
gunicorn==20.1.0
pytz
channels-redis
daphne==4.0